requirements.txt
asyncpg
Pillow
progressbar2
vk_api
vkcoin